requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
deprecation==2.1.0
docker==5.0.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
requests==2.25.1
selenium==3.141.0
six==1.16.0
testcontainers==3.4.0
toml==0.10.2
urllib3==1.26.4
webdriver-manager==3.4.1
websocket-client==0.59.0
wrapt==1.12.1
